Ref: Pipeline refactor & export split, make orchestration and export source-agnostic
