+++ scraper/pyproject.toml
@@ -53,7 +53,7 @@ Homepage = "https://github.com/openzim/gutenberg"
 Donate = "https://www.kiwix.org/en/support-us/"
 
 [project.scripts]
-gutenberg2zim = "gutenberg2zim.entrypoint:main"
+gutenberg2zim = "gutenberg2zim.cli:main"
 gutenberg2zim-validate-i18n = "gutenberg2zim.scripts.validate_i18n:main"
 
 [tool.hatch.version]
