+++ scraper/src/gutenberg2zim/core/i18n.py
