+++ scraper/src/gutenberg2zim/core/progress.py
@@ -2,7 +2,7 @@ import json
 import threading
 from math import ceil
 
-from schedule import every, run_pending
+from schedule import Scheduler
 
 from gutenberg2zim.constants import logger
 
@@ -15,22 +15,26 @@ class ScraperProgress:
         self._total = 1
         self._progress = 0
         self._stats_filename = stats_filename
+        # per-instance scheduler: jobs must not leak into schedule's global
+        # default scheduler (they would pin this instance and fire from
+        # unrelated instances' run_pending calls)
+        self._scheduler = Scheduler()
 
         if self._stats_filename:
             self.report_progress()
             # set a timer to report progress only every 10 seconds,
             # no need to do it more often
-            every(10).seconds.do(self.report_progress)
+            self._scheduler.every(10).seconds.do(self.report_progress)
 
     def increase_total(self, increment: int):
         with self._lock:
             self._total += increment
-            run_pending()
+            self._scheduler.run_pending()
 
     def increase_progress(self, increment: int = 1):
         with self._lock:
             self._progress += increment
-            run_pending()
+            self._scheduler.run_pending()
 
     def report_progress(self):
         if not self._stats_filename:
