+++ scraper/src/gutenberg2zim/core/rewriters/link_rewriter.py
@@ -1,0 +1,25 @@
+"""Generic relative link rewriting (moved from `sources/gutenberg/plugins.py`).
+
+Rewrites relative links to sibling HTML pages into `{item_id}_{page}` form
+so they resolve inside the ZIM. Source-agnostic: takes a plain item id.
+"""
+
+
+def replacement_link(item_id: int, url: str) -> str | None:
+    try:
+        urlp, anchor = url.rsplit("#", 1)
+    except ValueError:
+        urlp = url
+        anchor = None
+    if "/" in urlp:
+        return None
+
+    if len(urlp.strip()):
+        nurl = f"{item_id}_{urlp}"
+    else:
+        nurl = ""
+
+    if anchor is not None:
+        return "#".join([nurl, anchor])
+
+    return nurl
