+++ scraper/src/gutenberg2zim/core/schemas.py
