+++ scraper/src/gutenberg2zim/sources/gutenberg/plugins.py
@@ -1,30 +1,45 @@
-"""Gutenberg-specific HTML rewriting (moved from `gutenberg2zim.export`).
+"""Gutenberg-specific HTML rewriting and book export (from `export.py`).
 
 Contains everything needed to turn a Gutenberg HTML page into a static
 offline page: charset normalization, image path transformation, internal
 link rewriting, PG boilerplate ("*** START OF THE PROJECT GUTENBERG EBOOK")
 removal, and infobox injection. Exposed through the source-agnostic
-`RewriterPort` via `GutenbergHtmlRewriter`.
+`RewriterPort` via `GutenbergHtmlRewriter`. Also holds the per-book ZIM
+export (`export_book`) and EPUB optimization helpers.
 """
 
 import io
 import urllib.parse
-from dataclasses import asdict
+import warnings
+import zipfile
 from pathlib import Path
 
 import bs4
-from bs4 import BeautifulSoup, Tag
+from bs4 import BeautifulSoup, Tag, XMLParsedAsHTMLWarning
 from jinja2 import Environment, PackageLoader, select_autoescape
-from PIL.Image import open as pilopen
-from zimscraperlib.image.optimization import OptimizeWebpOptions
+from zimscraperlib.image.optimization import optimize_jpeg, optimize_png
 
-from gutenberg2zim.adapters import work_to_book
 from gutenberg2zim.constants import logger
 from gutenberg2zim.core.models import Work
 from gutenberg2zim.core.ports import RewriterPort
-from gutenberg2zim.utils import book_name_for_fs
+from gutenberg2zim.core.rewriters.image_rewriter import (
+    ImageProcessor,
+    rewrite_html_image_references,
+)
+from gutenberg2zim.core.rewriters.link_rewriter import replacement_link
+from gutenberg2zim.core.utils import (
+    UTF8,
+    archive_name_for,
+    article_name_for,
+    book_name_for_fs,
+    fname_for,
+)
+from gutenberg2zim.core.zim_assembler import ZimAssembler
+from gutenberg2zim.sources.gutenberg.adapters import work_to_book
+from gutenberg2zim.sources.gutenberg.downloader import download_book_cover
+from gutenberg2zim.sources.gutenberg.models import Book
 
-default_webp_options = asdict(OptimizeWebpOptions())
+warnings.filterwarnings("ignore", category=XMLParsedAsHTMLWarning)
 
 
 class HtmlRewriteError(RuntimeError):
@@ -98,33 +113,16 @@ def update_html_for_static(
         # Rewrite image references to use .webp extension for converted images
         # This also handles <link rel="icon"> tags
         # Only for regular HTML, not EPUB (EPUB images are not converted to WebP)
-        rewrite_html_image_references(soup, book)
+        rewrite_html_image_references(soup, f"Book {book.book_id}")
 
     # update all <a> links to internal HTML pages
     # should only apply to relative URLs to HTML files.
     # examples on #16816, #22889, #30021
-    def replacablement_link(book, url):
-        try:
-            urlp, anchor = url.rsplit("#", 1)
-        except ValueError:
-            urlp = url
-            anchor = None
-        if "/" in urlp:
-            return None
-
-        if len(urlp.strip()):
-            nurl = f"{book.book_id}_{urlp}"
-        else:
-            nurl = ""
-
-        if anchor is not None:
-            return "#".join([nurl, anchor])
-
-        return nurl
-
     if not epub:
         for link in soup.find_all("a"):
-            new_link = replacablement_link(book=book, url=link.attrs.get("href", ""))
+            new_link = replacement_link(
+                item_id=book.book_id, url=str(link.attrs.get("href", ""))
+            )
             if new_link is not None:
                 link.attrs["href"] = new_link
 
@@ -309,56 +307,31 @@ def update_html_for_static(
     return soup
 
 
-def rewrite_html_image_references(soup: BeautifulSoup, book) -> None:
-    """Rewrite HTML image references to use .webp extension for converted images."""
+def export_infobox_assets(assembler: ZimAssembler) -> None:
+    """Export infobox CSS, JS, and icon files to ZIM"""
+    templates_dir = Path(__file__).parent.parent.parent / "templates"
 
-    def rewrite_reference(element, attr, ref_type):
-        """Helper to rewrite a single reference attribute."""
-        if attr in element.attrs:
-            old_ref = element[attr]
-            new_ref = ImageProcessor.get_output_filename(old_ref)
-            if old_ref != new_ref:
-                element[attr] = new_ref
-                logger.debug(
-                    f"Book {book.book_id}: Rewrote {ref_type} {old_ref} -> {new_ref}"
-                )
+    assets = [
+        ("css/gutenberg-infobox.css", "css", "text/css"),
+        ("js/gutenberg-infobox.js", "js", "text/javascript"),
+        ("icons/info.svg", "icons", "image/svg+xml"),
+        ("icons/epub.svg", "icons", "image/svg+xml"),
+        ("icons/pdf.svg", "icons", "image/svg+xml"),
+        ("icons/scroll-up.svg", "icons", "image/svg+xml"),
+    ]
 
-    # Rewrite <img> tags
-    for img in soup.find_all("img"):
-        rewrite_reference(img, "src", "image")
-
-    # Rewrite <link rel="icon"> tags (for cover images in HTML head)
-    for link in soup.find_all("link", rel="icon"):
-        rewrite_reference(link, "href", "icon")
-
-
-class ImageProcessor:
-    """Centralized image processing logic for conversion decisions and operations."""
-
-    @staticmethod
-    def get_extension(filename: str) -> str:
-        """Extract lowercase file extension without the dot."""
-        return Path(filename).suffix[1:].lower()
-
-    @staticmethod
-    def should_convert_to_webp(filename: str) -> bool:
-        """Check if file should be converted to WebP (JPG, JPEG, PNG only)."""
-        return ImageProcessor.get_extension(filename) in ("jpg", "jpeg", "png")
-
-    @staticmethod
-    def get_output_filename(filename: str) -> str:
-        """Get output filename with .webp extension if file will be converted."""
-        if ImageProcessor.should_convert_to_webp(filename):
-            return str(Path(filename).with_suffix(".webp"))
-        return filename
-
-    @staticmethod
-    def optimize_image_content(file_content: bytes) -> bytes:
-        """Convert and optimize image content to WebP format."""
-        dst = io.BytesIO()
-        with pilopen(io.BytesIO(file_content)) as image:
-            image.save(dst, format="WEBP", **default_webp_options)
-        return dst.getvalue()
+    for zim_path, subdir, mimetype in assets:
+        file_path = templates_dir / subdir / Path(zim_path).name
+        if not file_path.exists():
+            logger.warning(f"Infobox asset not found: {file_path}")
+            continue
+        logger.debug(f"Adding {zim_path} to ZIM")
+        assembler.add_item_for(
+            path=zim_path,
+            fpath=file_path,
+            mimetype=mimetype,
+            is_front=False,
+        )
 
 
 class GutenbergHtmlRewriter(RewriterPort):
@@ -380,3 +353,302 @@ class GutenbergHtmlRewriter(RewriterPort):
             formats=self._formats,
         )
         return BeautifulSoup(str(result), "lxml")
+
+
+def export_book(
+    book: Book,
+    book_files: dict[str, bytes],
+    formats: list[str],
+    mirror_url: str,
+    assembler: ZimAssembler,
+    _zim_name: str,
+    *,
+    _title_search: bool,
+    _add_lcc_shelves: bool,
+):
+    """Export book to ZIM using in-memory content"""
+    handle_book_files(
+        book=book,
+        book_files=book_files,
+        formats=formats,
+        assembler=assembler,
+    )
+
+    # Handle cover image
+    cover_path = f"covers/{book.book_id}_cover_image.webp"
+
+    if book.html_cover_path:
+        # HTML has a cover image - create alias instead of storing duplicate
+        # Use alias (not redirect) since this is an image, not HTML with relative paths
+        logger.debug(
+            f"Using HTML cover for book #{book.book_id}: {book.html_cover_path}"
+        )
+        assembler.add_alias(
+            path=cover_path,
+            title="",
+            target=book.html_cover_path,
+        )
+    else:
+        # No HTML cover - download from mirror
+        cover_image = download_book_cover(mirror_url, book)
+
+        if cover_image:
+            logger.debug(f"Using downloaded cover for book #{book.book_id}")
+            # the mirror serves JPEG; convert to WebP to match cover_path/mimetype
+            cover_image = ImageProcessor.optimize_image_content(cover_image)
+            assembler.add_item_for(
+                path=cover_path,
+                content=cover_image,
+                mimetype="image/webp",
+                is_front=False,
+            )
+
+
+def handle_book_files(
+    book: Book,
+    book_files: dict[str, bytes],
+    formats: list[str],
+    assembler: ZimAssembler,
+):
+    """Handle book files from in-memory content and add to ZIM"""
+
+    # Find the main HTML file
+    main_html_filename = f"{book.book_id}.html"
+    html_content = None
+
+    if main_html_filename in book_files:
+        html_content = book_files[main_html_filename].decode("utf-8", errors="replace")
+
+    if html_content:
+        article_name = article_name_for(book)
+        new_html = update_html_for_static(
+            book=book, html_content=html_content, formats=formats
+        )
+
+        # Add the optimized HTML directly to ZIM
+        assembler.add_item_for(
+            path=article_name,
+            content=str(new_html),
+            mimetype="text/html",
+            is_front=False,
+            title=book.title,
+            auto_index=True,
+        )
+
+    # Handle other formats (epub, pdf)
+    other_filenames = []
+    for other_format in [
+        fmt for fmt in book.requested_formats(formats) if fmt != "html"
+    ]:
+        book_filename = fname_for(book, other_format)
+        if book_filename in book_files:
+            other_filenames.append(book_filename)
+            try:
+                archive_name = archive_name_for(book, other_format)
+                content = book_files[book_filename]
+                if other_format == "epub":
+                    content = optimize_epub_bytes(content, book)
+                assembler.add_item_for(
+                    path=archive_name,
+                    content=content,
+                    is_front=False,
+                )
+            except Exception as e:
+                logger.exception(e)
+                logger.error(f"\t\tException while handling {other_format}: {e}")
+                raise
+
+    # Process all associated files (images, companion HTML files, etc)
+    for filename, file_content in book_files.items():
+        # Skip the main HTML file as it's already processed
+        if filename == main_html_filename:
+            continue
+
+        # Skip files matching a specific format since they have already been processed
+        if filename in other_filenames:
+            continue
+
+        if filename.endswith((".html", ".htm")):
+            # Process companion HTML files
+            try:
+                html_str = file_content.decode("utf-8", errors="replace")
+                new_html = update_html_for_static(
+                    book=book, html_content=html_str, formats=formats
+                )
+                assembler.add_item_for(
+                    path=filename,
+                    content=str(new_html),
+                    mimetype="text/html",
+                    is_front=False,
+                )
+            except Exception as e:
+                logger.exception(e)
+                logger.error(f"\t\tException while handling companion HTML: {e}")
+        else:
+            # Add other files (images, etc) directly
+            try:
+                optimized_file_content = optimize_content(book, filename, file_content)
+                output_filename = ImageProcessor.get_output_filename(filename)
+
+                # Check if this is the cover image by comparing with transformed href
+                # Note: filename is already transformed (e.g., "1_cover.jpg")
+                # by download.py so we transform book._cover_href the same way
+                if book._cover_href:
+                    # Transform cover href same way we transform image paths
+                    expected_cover = transform_image_path(
+                        book.book_id, book._cover_href
+                    )
+                    expected_cover = ImageProcessor.get_output_filename(expected_cover)
+
+                    if output_filename == expected_cover:
+                        book.html_cover_path = output_filename
+                        logger.debug(
+                            f"Detected HTML cover for book #{book.book_id}: "
+                            f"{output_filename}"
+                        )
+
+                assembler.add_item_for(
+                    path=output_filename,
+                    content=optimized_file_content,
+                    is_front=False,
+                )
+            except Exception as e:
+                logger.exception(e)
+                logger.error(f"\t\tException while handling file {filename}: {e}")
+
+
+def optimize_content(book: Book, filename: str, file_content: bytes) -> bytes:
+    """Optimize file content, converting images to WebP when appropriate."""
+    # Convert JPG, PNG to WEBP for optimal file size
+    if ImageProcessor.should_convert_to_webp(filename):
+        return ImageProcessor.optimize_image_content(file_content)
+
+    # Keep WebP and GIF files as-is
+    ext = ImageProcessor.get_extension(filename)
+    if ext in ("webp", "gif"):
+        if ext == "gif":
+            logger.debug(
+                f"GIF file {filename} found in book {book.book_id} not optimized"
+            )
+        return file_content
+
+    # Do not optimize other file types
+    return file_content
+
+
+def _optimize_epub_jpeg(data: bytes) -> bytes:
+    """Optimize JPEG image in-memory for EPUB, keeping original format."""
+    dst = io.BytesIO()
+    optimize_jpeg(src=io.BytesIO(data), dst=dst)
+    return dst.getvalue()
+
+
+def _optimize_epub_png(data: bytes) -> bytes:
+    """Optimize PNG image in-memory for EPUB, keeping original format."""
+    dst = io.BytesIO()
+    optimize_png(src=io.BytesIO(data), dst=dst)
+    return dst.getvalue()
+
+
+def _process_epub_html(data: bytes, book: Book, *, is_xml: bool = False) -> bytes:
+    """Process HTML file from EPUB: remove Gutenberg markers and process content."""
+    html_str = data.decode("utf-8", errors="replace")
+    soup = update_html_for_static(
+        book=book, html_content=html_str, formats=[], epub=True, is_xml=is_xml
+    )
+    return str(soup).encode(UTF8)
+
+
+def _process_epub_ncx(data: bytes, book: Book | None = None) -> bytes:
+    """Process NCX navigation file: remove license section."""
+    ncx_str = data.decode("utf-8", errors="replace")
+    soup = BeautifulSoup(ncx_str, "lxml-xml")
+    pattern = "*** START: FULL LICENSE ***"
+    for tag in soup.find_all("text"):
+        if pattern in tag.text:
+            book_info = f"book {book.book_id}" if book else "unknown book"
+            logger.info(f"Found license section in NCX for {book_info}")
+            s = tag.parent.parent if tag.parent else None
+            if s is None:
+                logger.warning(f"Unexpected NCX structure for {book_info}")
+                break
+            # Collect siblings before decomposing (decompose breaks iteration)
+            siblings_to_remove = list(s.next_siblings)
+            s.decompose()
+            for sibling in siblings_to_remove:
+                if hasattr(sibling, "decompose"):  # Skip text nodes
+                    sibling.decompose()
+            break
+    return str(soup).encode(UTF8)
+
+
+def optimize_epub_bytes(epub_bytes: bytes, book: Book) -> bytes:
+    """Optimize EPUB in-memory: process HTML/NCX and optimize images without FS."""
+    src_buf = io.BytesIO(epub_bytes)
+    dst_buf = io.BytesIO()
+    original_size = len(epub_bytes)
+
+    with (
+        zipfile.ZipFile(src_buf, "r") as src_zf,
+        zipfile.ZipFile(dst_buf, "w", zipfile.ZIP_DEFLATED) as dst_zf,
+    ):
+        infos = src_zf.infolist()
+        mimetype_info = next(
+            (info for info in infos if info.filename == "mimetype"), None
+        )
+        if mimetype_info is None:
+            raise ValueError("EPUB is missing its mimetype entry")
+
+        # Write mimetype first, uncompressed, per EPUB spec
+        dst_zf.writestr(
+            "mimetype",
+            src_zf.read(mimetype_info),
+            compress_type=zipfile.ZIP_STORED,
+        )
+
+        for info in infos:
+            if info.filename == "mimetype":
+                continue
+
+            name = info.filename
+            data = src_zf.read(name)
+            suffix = Path(name).suffix.lower()
+
+            if suffix in (".jpg", ".jpeg"):
+                optimized_data = _optimize_epub_jpeg(data)
+                if len(optimized_data) < len(data):  # ignore bigger compressed version
+                    data = optimized_data
+            elif suffix == ".png":
+                optimized_data = _optimize_epub_png(data)
+                if len(optimized_data) < len(data):  # ignore bigger compressed version
+                    data = optimized_data
+            elif suffix in (".gif", ".webp"):
+                logger.warning(
+                    f"Unexpected {suffix} image in EPUB for book {book.book_id}: {name}"
+                )
+            elif suffix in (".htm", ".html", ".xhtml"):
+                data = _process_epub_html(data, book, is_xml=(suffix == ".xhtml"))
+            elif suffix == ".ncx":
+                data = _process_epub_ncx(data, book)
+
+            # copy metadata but force deflate: the source ZipInfo's
+            # compress_type would otherwise override the archive default
+            out_info = zipfile.ZipInfo(filename=info.filename, date_time=info.date_time)
+            out_info.external_attr = info.external_attr
+            out_info.compress_type = zipfile.ZIP_DEFLATED
+            dst_zf.writestr(out_info, data)
+
+    optimized_bytes = dst_buf.getvalue()
+    optimized_size = len(optimized_bytes)
+    if optimized_size > original_size:
+        logger.warning(
+            f"Optimized EPUB for book {book.book_id} is larger than original: "
+            f"{optimized_size} > {original_size} bytes"
+        )
+    else:
+        logger.debug(
+            f"Optimized EPUB for book {book.book_id}: "
+            f"{optimized_size} < {original_size} bytes"
+        )
+
+    return optimized_bytes
