--- scraper/src/gutenberg2zim/zim.py
@@ -1,204 +1,0 @@
-import datetime
-import re
-from html import escape
-from pathlib import Path
-
-from gutenberg2zim import i18n
-from gutenberg2zim.book_processor import process_all_books
-from gutenberg2zim.config import ScrapeConfig
-from gutenberg2zim.constants import logger
-from gutenberg2zim.core.work_store import WorkStore
-from gutenberg2zim.core.zim_assembler import ZimAssembler
-from gutenberg2zim.iso639 import ISO_MATRIX, ISO_MATRIX_REV, ZIM_LANGUAGES_MAP
-from gutenberg2zim.scraper_progress import ScraperProgress
-from gutenberg2zim.sources.gutenberg.catalog import CatalogEntry
-from gutenberg2zim.utils import get_zim_name
-
-
-def resolve_language(lang: str) -> list[str]:
-    """Helper to resolve one input language code to ZIM language code(s)."""
-    return [
-        zim_lang
-        for zim_lang in ZIM_LANGUAGES_MAP.get(
-            lang,
-            [ISO_MATRIX.get(lang, lang if lang in ISO_MATRIX_REV else None)],
-        )
-        if zim_lang
-    ]
-
-
-def get_zim_language_metadata(languages: list[str], books: list[CatalogEntry]):
-    unresolved = [lang for lang in languages if not resolve_language(lang)]
-    if unresolved:
-        raise ValueError(
-            f"Cannot resolve ZIM language metadata for: {', '.join(unresolved)}. "
-            "Use --zim-languages to override."
-        )
-    language_counts = {
-        zim_lang: sum(lang in book.languages for book in books)
-        for lang in languages
-        for zim_lang in resolve_language(lang)
-    }
-    return sorted(language_counts, key=lambda lang: language_counts[lang], reverse=True)
-
-
-def export_ui_dist(ui_dist: Path, title: str, assembler: ZimAssembler) -> None:
-    """Export Vue.js UI dist folder to ZIM file."""
-    if not ui_dist.exists():
-        raise FileNotFoundError(f"UI dist directory not found: {ui_dist}")
-
-    logger.info(f"Adding Vue.js UI files from {ui_dist}")
-    file_count = 0
-    for file in ui_dist.rglob("*"):
-        if file.is_dir():
-            continue
-        path = file.relative_to(ui_dist).as_posix()
-        logger.debug(f"Adding {path} to ZIM")
-
-        # Update index.html title
-        if path == "index.html":
-            html_content = file.read_text(encoding="utf-8")
-            new_html_content = re.sub(
-                r"(<title>)(.*?)(</title>)",
-                lambda match: f"{match.group(1)}{escape(title)}{match.group(3)}",
-                html_content,
-                flags=re.IGNORECASE,
-            )
-            assembler.add_item_for(
-                path=path,
-                content=new_html_content,
-                mimetype="text/html",
-                is_front=True,
-            )
-        else:
-            assembler.add_item_for(
-                path=path,
-                fpath=file,
-                is_front=False,
-            )
-        file_count += 1
-    logger.info(f"Successfully added {file_count} UI files to ZIM")
-
-
-def build_zimfile(
-    books: list[CatalogEntry],
-    config: ScrapeConfig,
-    work_store: WorkStore,
-    zim_file: str | None,
-    zim_name: str | None,
-    title: str | None,
-    description: str | None,
-    long_description: str | None,
-    zim_languages: list[str] | None,
-    publisher: str,
-    ui_dist: Path,
-    *,
-    primary_color: str | None = None,
-    secondary_color: str | None = None,
-    overwrite: bool,
-    is_selection: bool,
-    title_search: bool,
-    add_lcc_shelves: bool,
-    progress: ScraperProgress,
-    with_fulltext_index: bool,
-) -> None:
-    """Build ZIM file from the works collected in the work store"""
-    progress.increase_total(len(books))
-    output_folder = config.output_folder
-    mirror_url = config.mirror_url
-    concurrency = config.concurrency
-    languages = config.languages or []
-    formats = config.formats
-    debug = config.debug
-    iso_languages = [ISO_MATRIX.get(lang, lang) for lang in languages]
-
-    formats.sort()
-
-    metadata_lang = "mul" if len(iso_languages) > 1 else iso_languages[0]
-
-    metadata_locales_lang = ISO_MATRIX_REV.get(metadata_lang, metadata_lang)
-
-    i18n.change_locale(metadata_locales_lang)
-
-    title = title or i18n.t("metadata_defaults.title")
-    # check if user has description input otherwise assign default description
-    description = description or i18n.t(
-        "metadata_defaults.description",
-        f'All books in "{iso_languages[0]}" language from the first producer of free'
-        " Ebooks",
-    )
-
-    logger.info(f"\tWriting {metadata_lang} ZIM for {title}")
-
-    zim_name = zim_name or get_zim_name(languages, formats, is_selection)
-
-    if zim_file is None:
-        zim_file = "{}_{}.zim".format(
-            zim_name,
-            datetime.datetime.now().strftime("%Y-%m"),  # noqa: DTZ005
-        )
-        zim_path = output_folder / zim_file
-    else:
-        zim_path = Path(zim_file)
-        if not zim_path.is_absolute() and zim_path.parent == Path("."):
-            # Just a filename, put it in output_folder
-            zim_path = output_folder / zim_path
-
-    # Ensure the output folder exists before creating the ZIM
-    zim_path.parent.mkdir(parents=True, exist_ok=True)
-
-    if zim_path.exists() and not overwrite:
-        logger.info(f"ZIM file `{zim_file}` already exist.")
-        return
-    elif zim_path.exists():
-        logger.info(f"Removing existing ZIM file {zim_file}")
-        zim_path.unlink(missing_ok=True)
-
-    assembler = ZimAssembler(
-        filename=zim_path,
-        language=zim_languages or get_zim_language_metadata(languages, books),
-        title=title,
-        description=description,
-        long_description=long_description,
-        name=zim_name,
-        publisher=publisher,
-        source_creator="gutenberg.org",
-        tags="_category:gutenberg;gutenberg",
-        with_fulltext_index=with_fulltext_index,
-        debug=debug,
-    )
-
-    assembler.start()
-
-    try:
-        process_all_books(
-            book_ids=[book.book_id for book in books],
-            zim_name=zim_name,
-            mirror_url=mirror_url,
-            concurrency=concurrency,
-            _languages=languages,
-            formats=formats,
-            progress=progress,
-            work_store=work_store,
-            assembler=assembler,
-            title_search=title_search,
-            add_lcc_shelves=add_lcc_shelves,
-            title=title,
-            description=description,
-            primary_color=primary_color,
-            secondary_color=secondary_color,
-        )
-
-        # Export Vue.js UI dist folder
-        export_ui_dist(ui_dist, title, assembler)
-
-    except KeyboardInterrupt:
-        assembler.can_finish = False
-        logger.error("KeyboardInterrupt, exiting.")
-        raise
-    except Exception:
-        assembler.can_finish = False
-        logger.exception("Interrupting process due to error")
-        raise
-    else:
-        assembler.finish()
